Wiele zestawów pytań na jednego usera

+++ Controllers/QuestionsController.cs
@@ -4,12 +4,12 @@ using System.Diagnostics;
 using System.Linq;
 using System.Web.Mvc;
 using QuizR.Models;
+using System.Data.Entity;
 
 namespace QuizR.Controllers
 {
     public class QuestionsController : Controller
     {
-
         public ApplicationDbContext db { get; set; }
         public UserManager<ApplicationUser> UserManager { get; set; }
 
@@ -20,30 +20,26 @@ namespace QuizR.Controllers
         }
 
 
-        // GET: Questions
+        // GET: Questions/5 (id zestawu)
         [Authorize]
-        public ActionResult Index()
+        public ActionResult Manage(int id)
         {
+            //sprawdzanie czy jest wlascicielem
             var user = UserManager.FindById(User.Identity.GetUserId());
-            var set = db.QuestionSets.FirstOrDefault(s => s.Owner.Id == user.Id);
-            if(set == null)
-            {
-                set = new QuestionSet() { Owner = user };
-                db.QuestionSets.Add(set);
-                db.SaveChanges();
-            }
+            var set = db.QuestionSets.Find(id);
+            if (set == null || !set.Owner.Equals(user)) return View("Error");
+
             return View(set);
         }
 
-        // GET: Questions/Details/5
-        public ActionResult Details(int id)
-        {
-            return View();
-        }
-
         // GET: Questions/Create
-        public ActionResult Create()
+        public ActionResult Create(int setID)
         {
+            //sprawdzanie czy jest wlascicielem
+            var user = UserManager.FindById(User.Identity.GetUserId());
+            var set = db.QuestionSets.Find(setID);
+            if (set == null || !set.Owner.Equals(user)) return View("Error");
+            ViewBag.SetID = setID;
             return View();
         }
 
@@ -53,22 +49,28 @@ namespace QuizR.Controllers
         {
             try
             {
+                //parsowanie danych z formularza
                 string content = collection["Content"];
                 string[] ans = new string[4];
                 ans[0] = collection["Aans"];
                 ans[1] = collection["Bans"];
                 ans[2] = collection["Cans"];
                 ans[3] = collection["Dans"];
-
                 string correct = collection["Correct"];
+                var setID = int.Parse(collection["setID"]);
 
+                var set = db.QuestionSets.Find(setID);
+
+                //sprawdzanie czy jest wlascicielem
                 var user = UserManager.FindById(User.Identity.GetUserId());
-                var set = db.QuestionSets.FirstOrDefault(s => s.Owner.Id == user.Id);
-                var q = new Question() { Content = content, Aans = ans[0], Bans = ans[1], Cans = ans[2], Dans = ans[3], Correct = correct };
+                if (set == null || !set.Owner.Equals(user)) return View("Error");
+
+
+                var q = new Question() { Owner = user, Set = set, Content = content, Aans = ans[0], Bans = ans[1], Cans = ans[2], Dans = ans[3], Correct = correct };
                 set?.Questions.Add(q);
                 db.SaveChanges();
 
-                return RedirectToAction("Index");
+                return RedirectToAction("Manage", new { id = setID });
             }
             catch
             {
@@ -79,7 +81,16 @@ namespace QuizR.Controllers
         // GET: Questions/Edit/5
         public ActionResult Edit(int id)
         {
-            return View();
+            var q = db.Questions
+                    .Include(r => r.Owner)
+                    .Include(r => r.Set)
+                    .SingleOrDefault(r => r.ID == id);
+
+            //sprawdzanie czy jest wlascicielem
+            var user = UserManager.FindById(User.Identity.GetUserId());
+            if (q == null || !q.Owner.Equals(user)) return View("Error");
+
+            return View(q);
         }
 
         // POST: Questions/Edit/5
@@ -88,9 +99,24 @@ namespace QuizR.Controllers
         {
             try
             {
-                // TODO: Add update logic here
+                var q = db.Questions
+                    .Include(r => r.Owner)
+                    .Include(r => r.Set)
+                    .SingleOrDefault(r => r.ID == id);
 
-                return RedirectToAction("Index");
+                //sprawdzanie czy jest wlascicielem
+                var user = UserManager.FindById(User.Identity.GetUserId());
+                if (q == null || !q.Owner.Equals(user)) return View("Error");
+
+                q.Content = collection["Content"];
+                q.Aans = collection["Aans"];
+                q.Bans = collection["Bans"];
+                q.Cans = collection["Cans"];
+                q.Dans = collection["Dans"];
+                q.Correct = collection["Correct"];
+
+                db.SaveChanges();
+                return RedirectToAction("Manage", new { id = q.Set.ID });
             }
             catch
             {
@@ -101,7 +127,16 @@ namespace QuizR.Controllers
         // GET: Questions/Delete/5
         public ActionResult Delete(int id)
         {
-            var q = db.Questions.FirstOrDefault(r => r.ID == id);
+            var q = db.Questions
+                    .Include(r => r.Owner)
+                    .Include(r => r.Set)
+                    .SingleOrDefault(r => r.ID == id);
+
+
+            //sprawdzanie czy jest wlascicielem
+            var user = UserManager.FindById(User.Identity.GetUserId());
+            if (q == null || !q.Owner.Equals(user)) return View("Error");
+
             return View(q);
         }
 
@@ -109,9 +144,20 @@ namespace QuizR.Controllers
         [HttpPost]
         public ActionResult Delete(int id, FormCollection collection)
         {
+            var q = db.Questions
+                    .Include(r => r.Owner)
+                    .Include(r => r.Set)
+                    .SingleOrDefault(r => r.ID == id);
+            
+
+            //sprawdzanie czy jest wlascicielem
+            var user = UserManager.FindById(User.Identity.GetUserId());
+            if (q == null || !q.Owner.Equals(user)) return View("Error");
+            
+            var setID = q.Set.ID;
+            
             try
             {
-                var q = db.Questions.FirstOrDefault(r => r.ID == id);
                 db.Questions.Remove(q);
                 db.SaveChanges();
             }
@@ -120,7 +166,7 @@ namespace QuizR.Controllers
                 Debug.WriteLine("Nie udalo sie usunac!");
             }
 
-            return RedirectToAction("Index");
+            return RedirectToAction("Manage", new { id = setID });
         }
     }
 }
